WIP: tests, fix bugs and refactoring
