+++ compiler/Compiler.py
@@ -10,11 +10,7 @@ from aiofile import async_open
 from compiler.types.ide_types import SupportedCompilers
 from compiler.config import LIBRARY_PATH, BUILD_DIRECTORY
 from compiler.types.inner_types import CommandResult, BuildFile, File
-
-
-def get_file_extension(suffixes: List[str]) -> str:
-    """Create string extension from Path.suffixes."""
-    return ''.join(suffixes).replace('.', '', 1)
+from compiler.utils import get_file_extension, get_filename
 
 
 async def get_build_files(
@@ -22,7 +18,9 @@ async def get_build_files(
     """
     Get all files from 'build' directory in project path direcrory.
 
-    Create build directory if doesn't exist..
+    Create build directory if doesn't exist.
+
+    Function join project_path and ./build.
     """
     project_build_directory = project_path.joinpath('./build')
     await project_path.mkdir(exist_ok=True)
@@ -32,8 +30,8 @@ async def get_build_files(
         async with async_open(path, 'rb') as f:
             file_data = await f.read()
             extension = get_file_extension(path.suffixes)
-            filename = str(path.relative_to(
-                project_build_directory)).split('.')[0]
+            filename = get_filename(str(path.relative_to(
+                project_build_directory)))
             yield BuildFile(filename=filename,
                             extension=extension,
                             fileContent=file_data)
@@ -222,12 +220,13 @@ class Compiler:
 
     @staticmethod
     async def include_source_files(platform_id: str,
+                                   platform_version: str,
                                    libraries: Set[str],
                                    target_directory: str
                                    ) -> None:
         """Include source files from platform's \
             library directory to target directory."""
-        path = os.path.join(LIBRARY_PATH, f'{platform_id}/')
+        path = os.path.join(LIBRARY_PATH, f'{platform_id}/{platform_version}/')
         path_to_libs = set([os.path.join(path, library)
                            for library in libraries])
         await asyncio.create_subprocess_exec('cp',
@@ -236,7 +235,7 @@ class Compiler:
                                              cwd=BUILD_DIRECTORY)
 
     @staticmethod
-    async def includeLibraryFiles(
+    async def include_library_files(
             libraries: Set[str],
             target_directory: str,
             extension: str,
